reorganize server
