+++ database/helpers/availabilitiesHelper.js
