+++ database/helpers/eventsHelper.js
