+++ database/helpers/hoursOfOperationHelper.js
